Move serialization capabilities of circular buffer to separate, derived class
