+++ open_spiel/utils/CMakeLists.txt
@@ -17,6 +17,7 @@ add_library (utils OBJECT
   random.cc
   run_python.h
   run_python.cc
+  serializable_circular_buffer.h
   serialization.h
   stats.h
   tensor_view.h
@@ -67,6 +68,13 @@ if (OPEN_SPIEL_BUILD_WITH_PYTHON)
                  $<TARGET_OBJECTS:tests>)
   add_test(run_python_test run_python_test
            --python_command=${Python3_EXECUTABLE})
+endif()
+
+if (OPEN_SPIEL_BUILD_WITH_LIBNOP)
+  add_executable(serializable_circular_buffer_test
+                 serializable_circular_buffer_test.cc ${OPEN_SPIEL_OBJECTS}
+                 $<TARGET_OBJECTS:tests>)
+  add_test(serializable_circular_buffer_test serializable_circular_buffer_test)
 endif()
 
 add_executable(stats_test stats_test.cc ${OPEN_SPIEL_OBJECTS}
